do not try to overwrite existing hashed source files

+++ conda_build/source.py
@@ -101,7 +101,8 @@ def download_to_cache(cache_folder, recipe_path, source_dict):
         if not hashed:
             hashed = hashsum_file(path, 'sha256')
         dest_path = append_hash_to_fn(path, hashed)
-        os.rename(path, dest_path)
+        if not os.path.isfile(dest_path):
+            os.rename(path, dest_path)
         path = dest_path
 
     return path, unhashed_fn
